requirements.txt
flask==2.3.3
tensorflow==2.13.1
Pillow==10.0.1
werkzeug==2.3.7
numpy==1.24.3
